+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to
 
 ## Added
 
+- ✨(backend) limit link reach/role select options depending on ancestors #645
 - ✨(backend) add new "descendants" action to document API endpoint #645
 - ✨(backend) new "tree" action on document detail endpoint #645
 - ✨(backend) allow forcing page size within limits #645